+++ src/Resources/UserPasswordResource.php
@@ -51,14 +51,14 @@ class UserPasswordResource extends BaseRestResource
     {
         $oldPassword = $this->getPayloadData('old_password');
         $newPassword = $this->getPayloadData('new_password');
+        $login = $this->request->getParameterAsBool('login');
 
         if (!empty($oldPassword) && Session::isAuthenticated()) {
             $user = Session::user();
 
-            return static::changePassword($user, $oldPassword, $newPassword);
+            return static::changePassword($user, $oldPassword, $newPassword, $login);
         }
 
-        $login = $this->request->getParameterAsBool('login');
         $email = $this->getPayloadData('email');
         $username = $this->getPayloadData('username');
         $code = $this->getPayloadData('code');
@@ -89,7 +89,7 @@ class UserPasswordResource extends BaseRestResource
             return static::changePasswordBySecurityAnswer($email, $answer, $newPassword, $login);
         }
 
-        return false;
+        throw new BadRequestException('Not enough information provided to change password.');
     }
 
     /**
@@ -263,13 +263,14 @@ class UserPasswordResource extends BaseRestResource
      * @param User   $user
      * @param string $old
      * @param string $new
+     * @param bool   $login
      *
      * @return array
      * @throws BadRequestException
      * @throws InternalServerErrorException
      * @throws NotFoundException
      */
-    protected static function changePassword(User $user, $old, $new)
+    protected static function changePassword(User $user, $old, $new, $login = true)
     {
         static::isAllowed($user);
 
@@ -298,6 +299,12 @@ class UserPasswordResource extends BaseRestResource
         try {
             $user->password = $new;
             $user->save();
+
+            if ($login) {
+                static::userLogin($user->email, $new);
+
+                return ['success' => true, 'session_token' => Session::getSessionToken()];
+            }
 
             return array('success' => true);
         } catch (\Exception $ex) {