+++ tests/SystemServiceTest.php
@@ -42,6 +42,6 @@ class SystemServiceTest extends \DreamFactory\Core\Testing\TestCase
         $this->assertEquals('system', Arr::get($content, 'name'));
         $this->assertEquals('System Management', Arr::get($content, 'label'));
         $this->assertEquals(1, Arr::get($content, 'id'));
-        $this->assertEquals(3, count($content));
+        $this->assertEquals(4, count($content));
     }
 }